% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateAmplicons.R
\name{annotateAmplicons}
\alias{annotateAmplicons}
\title{Function: annotateAmplicons
-------------------------------
This function takes a SingleCellExperiment object as input and annotates the stored amplicons.}
\usage{
annotateAmplicons(sce, known.canon)
}
\arguments{
\item{sce}{SingleCellExperiment object containing the single-cell data.}
}
\value{
A dataframe containing annotated amplicons.
}
\description{
Function: annotateAmplicons
-------------------------------
This function takes a SingleCellExperiment object as input and annotates the stored amplicons.
}
\examples{
\dontrun{
# Assume `sce` is a SingleCellExperiment object with a 'counts' assay
h5_file_path <- system.file("extdata", "demo.h5", package = "scafari")
sce <- h5ToSce(h5_file_path)
annotated <- annotateAmplicons(sce)
}
}
