% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateVariants.R
\name{annotateVariants}
\alias{annotateVariants}
\title{Function: annotateVariants
-------------------------------
This function takes a SingleCellExperiment object as input and performs variant annotation.}
\usage{
annotateVariants(sce, shiny = FALSE)
}
\arguments{
\item{sce}{SingleCellExperiment object containing the single-cell data to be annotated.}

\item{shiny}{A logical flag indicating whether the function is being run in a Shiny application
context. Default is FALSE.}
}
\value{
The function returns an annotated SingleCellExperiment object.
}
\description{
Function: annotateVariants
-------------------------------
This function takes a SingleCellExperiment object as input and performs variant annotation.
}
\examples{

\dontrun{
# Assume `sce` is a SingleCellExperiment object with variants in altExp()
sce <- annotateVariants(sce, shiny = FALSE)
}

}
\references{
https://missionbio.github.io/mosaic/, https://github.com/rachelgriffard/optima
}
