% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_MBAPI}
\alias{check_MBAPI}
\title{Check if MissionBio is accessible}
\usage{
check_MBAPI()
}
\value{
boolean if MissionBio API is currently available
}
\description{
Check if MissionBio is accessible
}
\keyword{internal}
