% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVariants.R
\name{filterVariants}
\alias{filterVariants}
\title{Function: filterVariants
-------------------------------
This function takes a SingleCellExperiment object as input and performs variant filtering}
\usage{
filterVariants(
  depth.threshold = numeric(),
  genotype.quality.threshold = numeric(),
  vaf.ref = numeric(),
  vaf.het = numeric(),
  vaf.hom = numeric(),
  min.cell = numeric(),
  min.mut.cell = numeric(),
  se.var,
  sce,
  shiny = FALSE
)
}
\arguments{
\item{depth.threshold}{A numeric value specifying the minimum read depth required.}

\item{genotype.quality.threshold}{A numeric value specifying the minimum genotype quality score.}

\item{vaf.ref}{A numeric value specifying the variant allele frequency threshold for wild-type alleles.}

\item{vaf.het}{A numeric value specifying the variant allele frequency threshold for heterozygous variants.}

\item{vaf.hom}{A numeric value specifying the variant allele frequency threshold for homozygous variants.}

\item{min.cell}{A numeric value indicating the minimum number of cells with a genotype other than "missing".}

\item{min.mut.cell}{A numeric value indicating the minimum number of mutated (genotype either "homozygous" or "heterozygous") cells.}

\item{se.var}{The SummarizedExperiment object containing variant data which will be filtered.}

\item{sce}{SingleCellExperiment object containing the single-cell data TODO.}

\item{shiny}{A logical flag indicating whether the function is being run in a Shiny application
context. Default is FALSE.}
}
\value{
A list containing the following elements:
\describe{
  \item{vaf.matrix.filtered}{Variant allele frequencies after filtering.}
  \item{genotype.matrix.filtered}{Genotype information after filtering.}
  \item{read.counts.df.norm.filtered}{Normalized read counts for variants retained after filtering.}
  \item{variant.ids.filtered}{A vector of the variant IDs that were retained after filtering.}
  \item{genoqual.matrix.filtered}{Genotype qualities for variants retained after filtering.}
  \item{cells.keep}{A vector of cell identifiers for those cells retained after filtering.}
}
}
\description{
Function: filterVariants
-------------------------------
This function takes a SingleCellExperiment object as input and performs variant filtering
}
\references{
https://missionbio.github.io/mosaic/, https://github.com/rachelgriffard/optima
}
