% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchScafariShiny.R
\name{launchScafariShiny}
\alias{launchScafariShiny}
\title{Launch the Scafari Shiny App}
\usage{
launchScafariShiny()
}
\value{
shiny app
}
\description{
Launch the Scafari Shiny App
}
\examples{
launchScafariShiny()

}
