% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotElbow.R
\name{plotElbow}
\alias{plotElbow}
\title{This function takes a SingleCellExperiment object and variants of interest as input and plots an elbow plot to perform k-means later.}
\usage{
plotElbow(sce, variants.of.interest)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing the relevant data.}

\item{variants.of.interest}{A vector specifying the variants of interest.}
}
\description{
This function takes a SingleCellExperiment object and variants of interest as input and plots an elbow plot to perform k-means later.
}
