% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenotypeDistributionPie.R
\name{plotGenotypeDistributionPie}
\alias{plotGenotypeDistributionPie}
\title{Plot genotype distribution pie}
\usage{
plotGenotypeDistributionPie(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing the relevant data.}
}
\value{
ggplot object visualizing the genotype distribution in a pie chart.

\dontrun{
# Assume `sce` is a SingleCellExperiment object with variants in altExp()
pie_chart <- plotGenotypeDistributionPie(sce)
print(pie_chart)
}
}
\description{
This function generates a plot to visualize the genotype distribution.
}
