% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterVariantSelection.R
\name{clusterVariantSelection}
\alias{clusterVariantSelection}
\title{Function: clusterVariantSelection
-------------------------------
This function takes selected variants and performs clustering on them.}
\usage{
clusterVariantSelection(sce, variants.of.interest, n.clust)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing the single-cell data on which clustering will be performed.}

\item{variants.of.interest}{A vector or list specifying the variants of interest to be selected for clustering.}

\item{n.clust}{An integer specifying the number of clusters.}
}
\value{
A list with k-means results and a ggplot-object.
}
\description{
Function: clusterVariantSelection
-------------------------------
This function takes selected variants and performs clustering on them.
}
\examples{
# Assume `sce` is a SingleCellExperiment object with variants in altExp()
sce_filtered <- readRDS(system.file("extdata", "sce_filtered.rds", package = "scafari"))
clusterplot <- clusterVariantSelection(
  sce = sce_filtered,
  variants.of.interest = c(
    "FLT3:chr13:28610183:A/G",
    "KIT:chr4:55599436:T/C",
    "TP53:chr17:7577427:G/A",
    "TET2:chr4:106158216:G/A"
  ),
  n.clust = 4
)

}
