% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchScafariShiny.R
\name{launchScafariShiny}
\alias{launchScafariShiny}
\title{Launch Scafari Shiny Application}
\usage{
launchScafariShiny()
}
\value{
shiny app
}
\description{
Launches the Scafari Shiny application for data visualization.
}
\examples{
if (interactive()) {
  launchScafariShiny()
}

}
