% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterGenotype.R
\name{plotClusterGenotype}
\alias{plotClusterGenotype}
\title{Plot Genotype Clusters}
\usage{
plotClusterGenotype(sce, variants.of.interest, gg.clust)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing the relevant data.}

\item{variants.of.interest}{A vector specifying the variants of interest.}

\item{gg.clust}{An object containing clustering information.}
}
\value{
A ggplot object that visually represents the clustering of genotypes based on the specified variants and clustering information.
}
\description{
This function generates a plot to visualize genotype in clusters based on selected variants of interest.
}
\examples{
# Assume `sce` is a SingleCellExperiment object with variants in altExp() and clusterplot is the output of clusterVariantSleection().
plotClusterGenotype(sce = sce_filtered, 
               variants.of.interest = c("FLT3:chr13:28610183:A/G",
                                        "KIT:chr4:55599436:T/C",
                                        "TP53:chr17:7577427:G/A",
                                        "TET2:chr4:106158216:G/A"), 
               gg.clust = clusterplot$clusterplot)

}
